refactor(*): use a single class to update monthly totals
Currently, transacting amounts into a month's balance has a few
problems:
- it's not atomic, meaning you must manually ensure that both monthly
budget and month records are updated
- has the side effect of creating new months or monthly budgets instead
of receiving them as parameters

To address that a new class is proposed: it solves the above problems by
using dep. injection - that is, both monthly budget and month - should
already exist and be passed as argument; and by considering the amount
update as a single, atomic operation instead of two different things.