refactor: rename account related columns on transactions
Since transactions are now seen as atomic patches to an account balance,
it does not make sense to have a single transaction object have both
origin and destination.

Instead, transactions have a single relationship with accounts and may
be
linked to other transactions if they need to represent a movement
comprising of two separate, atomic patches.